Two Factor Authentication
Fuckity, fuckity fuck fuck!

I originaly didn't know how I was going to implement the work-flow for
2FA so, I decided to make it look like GitHub.... If user has 2fa
enabled, they are taken to `/session/two-factor`, they enter their token
and are logged in (if it's correct).

I was going to encrypt/decrypt the shared key but... It's too much
bloody work ATM. I just want to get this app working.