User search!
`/users/` is now a search form to allow for quick search of users. It's
currently not limited (so, if we have 100 "Havoc" and search for "havoc"
we'll get 100 results).

The search function allows us to search the user's email, username or
ID. Gives us a fair amount of flexibility. We can probably change it so
it only searches for emails or IDs... But, I feel searching all of them
is good enough.