Added more to `/profile`
Changed up the HTML for the profile page.. Just to show what information is available and how to create a dynamic page based on said data.
I've also added Socket.IO to the dependencies, going to need it for a future update (chats, yey :D).

I've changed all references to `/account`to `/profile` because "profile" just seems more professional.