Added a "nice" event page.
When users scroll to the bottom of `/events` the website will automatically send a request to the event api for the next lot of events to show to the user.

I just need to make the whole thing a bit cleaner. E.g. loading GIF when sending the request for more events.