More authentication methods & Adde profileImage
Added a couple of functions to allow Mods & Admins access to certain routes (or just Admins).

I've also added a virtual field in the user schema called "profileImage". When you get it, it returns either the local image location (e.g. /images/{ID}) or a imaged based on the user's ID.. Note: Robohash went down :( so, I'm now using flathash.

I've also added an annoying bug to allow you to return to the previous page when you log in (yey).