Easy algorithm upgrading.
If the algorithms being used for security reasons get cracked or are found to have a massive insecurity, it can be easily updated by changing the values of "hashAlgo" and "cipherAlgo" in `config.json`. 

Honestly, I't feels like i've done a massive botch job to get this to work.... But it works nonetheless.

I've also changed the profile HTML to stop the password prompt (it's stored in session).

Some UUIDs have been modified to make it slightly easier for users to enter them (made shorter and removed hyphens (-) )

The profile JS now uses `User#profileImage` from the DB.