Event creation & Security update
Added "/events/create" route and a temp form for creating events.
I've also gone mad with making sure the input we get from the user is safe before it goes anywhere near the database or is shown to other users. I hope I got everything, I might not have and we'll porbably get hacked in the next week.